Issue: pgp is undefined in utility functions
This is happening because of the order Node is initializing the
database object, and therefore the repos of associated with the
database. Because I am attempting to import the util function
into the repo (and pgp, by extension) pgp is not yet defined
in the utility.

There are two solutions that I can think of. First, use pgp as
a parameter for the utility functions and pass this in when
using the functions. I don't think that this is ideal. Second,
require the pg-promise library (thus pgp) in my utility functions.
This will provide an uninitialized pgp object to be available in
my utility functions.