Refactor: working on refactoring repo composition
The biometrics repo & service currently work as one would
expect. The order of operations is:
resolver > service > repo

The delegation of duties are as follows, repo takes care
of formatting queries. The service runs the queries. The
resolver calls the query runner from the service.

Should the service extend the repo? IE dataloaders, custom SQL?