Fix: check null before calling loader
check https://github.com/graphql/dataloader/issues/98. Must check
for null values before calling the loader, otherwise the dataloader
returns an error: (must call loader with value, received null).

The fix is simple, but not clean. I think the best way to do this
will to create a static DataLoader function instance outside of
the repo, or in the services. This function will return a DataLoader.
The first thing it will do is check if the key is null, if it is,
then it will return null. I think. Reread the issue for more info.